Adds a 'claim' tag for failed builds.
Exactly like the 'blame' tag, add a 'claim' tag to unsuccessful builds.
This tag will be shown when somebody 'claimed' the build (using the
claim plugin).
It is possible to choose to display one, the other, or both the 'claim'
and 'blame' tags through the css display: property. If your process
works better with blame, set the .claim class to 'display: none' and it won't
show.